remove old code, fix types import
